requirements.txt
streamlit
pandas
numpy
scikit-learn
bs4
plotly-express
requests
pdfminer.six
docx2txt
textract
nltk
pathlib
sentence_transformers